Auto authenticate user as guest in rest router if no user is authenticated
