+++ src/modules/login/router.js
@@ -1,57 +1,17 @@
-const config = require('../../../config');
-const jwt = require('jsonwebtoken');
 const userMiddleware = require('../../middlewares/user');
 const authMiddleware = require('../../middlewares/auth');
+const autoLoginMiddleware = require('../../middlewares/auto-login');
 const parametersMiddleware = require('../../middlewares/parameters');
 const uuid = require('../../uuid');
 const _ = require('lodash');
 const q = require('./query');
-const db = require('../../db');
 const Joi = require('../../joi');
-
-const GUEST_KEY = 'cad8ea0d-f95e-43c1-b162-0704bfc1d3f6';
+const auth = require('./auth');
 
 const UserType = {
     USER: 'user',
     GUEST: 'guest',
 };
-
-/**
- * @param {Object} payload
- * @param {String} payload.key Key from database in case `type` is `user` or random in case `type` is `guest`
- * @param {String} payload.realKey Key from database
- * @param {String} payload.type `guest` or `user`
- *
- * @returns {Object} Payload
- */
-function tokenPayload({key, type, realKey}) {
-    return {key, type, realKey};
-}
-
-/**
- * @param {Object} payload
- * @param {String} payload.key Key from database in case `type` is `user` or random in case `type` is `guest`
- * @param {String} payload.realKey Key from database
- * @param {String} payload.type `guest` or `user`
- *
- * @returns Promise
- */
-function createAuthToken(payload) {
-    return new Promise((resolve, reject) => {
-        jwt.sign(
-            payload,
-            config.jwt.secret,
-            {expiresIn: config.jwt.expiresIn},
-            function (err, token) {
-                if (err == null) {
-                    return resolve(token);
-                }
-
-                reject(err);
-            }
-        );
-    });
-}
 
 function formatPermissions(permissions, plan) {
     const permissionsByResourceType = _.groupBy(
@@ -98,23 +58,6 @@ async function getLoginInfo(user, token, plan) {
     };
 }
 
-async function autoLogin(request, response, next) {
-    if (request.user != null) {
-        return next();
-    }
-
-    const user = {
-        key: uuid.generate(),
-        type: UserType.GUEST,
-        realKey: GUEST_KEY,
-    };
-    const token = await createAuthToken(tokenPayload(user));
-
-    request.user = user;
-    request.authToken = token;
-    next();
-}
-
 const LoginBodySchema = Joi.object().meta({className: 'Login'}).keys({
     username: Joi.string().required(),
     password: Joi.string().required(),
@@ -158,8 +101,8 @@ module.exports = (plan) => [
                     return;
                 }
 
-                const token = await createAuthToken(
-                    tokenPayload({
+                const token = await auth.createAuthToken(
+                    auth.tokenPayload({
                         ...user,
                         ...{type: UserType.USER, realKey: user.key},
                     })
@@ -197,7 +140,7 @@ module.exports = (plan) => [
             tags: ['login'],
         },
         responses: {200: {}},
-        middlewares: [userMiddleware, autoLogin, authMiddleware],
+        middlewares: [userMiddleware, autoLoginMiddleware, authMiddleware],
         handler: async function (request, response) {
             response
                 .status(200)
